catalogs services added permission check (create, delete)

+++ catalogs_services/api/routers/router_product.py
@@ -1,10 +1,8 @@
-import json
-from decimal import Decimal
-from typing import Annotated, List
+from typing import Annotated
 from uuid import UUID
 
 from sqlalchemy.ext.asyncio import AsyncSession
-from fastapi import APIRouter, Depends, status, Request, HTTPException
+from fastapi import APIRouter, Depends, status, Request
 from aiokafka import AIOKafkaProducer
 
 from api.CRUD.crud_products import (
@@ -15,7 +13,9 @@ from api.CRUD.crud_products import (
 )
 
 from api.dependencies.kafka_state import get_producer
-from api.dependencies.service_httpx import permission_product
+
+from api.dependencies.validate_premission import check_permission
+from api.dependencies.validate_price import validate_price_decimal
 from core.model import helper_db
 from core.schemas.schema_product import ProductRead, ProductCreate, ProductUpdate
 from core.model import Product
@@ -58,12 +58,10 @@ async def create_new_product(
     request: Request,
 ) -> Product:
     permission_codes = request.state.user.get("permission")
-    if "product.create" not in permission_codes:
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="No access for this operation",
-        )
-
+    await check_permission(
+        permission_codes=permission_codes,
+        product_code="product.create",
+    )
     product = await create_product(session=session, data_product=data_product)
     return product
 
@@ -74,6 +72,7 @@ async def create_new_product(
     status_code=status.HTTP_200_OK,
 )
 async def update_product_by_id(
+    request: Request,
     data_update: ProductUpdate,
     session: Annotated[AsyncSession, Depends(helper_db.session_getter)],
     producer: Annotated[AIOKafkaProducer, Depends(get_producer)],
@@ -106,6 +105,7 @@ async def update_product_by_id(
     status_code=status.HTTP_200_OK,
 )
 async def update_product_by_id_partial(
+    request: Request,
     data_update: ProductUpdate,
     session: Annotated[AsyncSession, Depends(helper_db.session_getter)],
     producer: Annotated[AIOKafkaProducer, Depends(get_producer)],
@@ -125,7 +125,10 @@ async def update_product_by_id_partial(
     )
 
     product_update = await update_product(
-        session=session, data_update=data_update, product=product, partial=True
+        session=session,
+        data_update=data_update,
+        product=product,
+        partial=True,
     )
     return ProductRead.model_validate(product_update)
 
@@ -135,8 +138,15 @@ async def update_product_by_id_partial(
     status_code=status.HTTP_204_NO_CONTENT,
 )
 async def delete_product_by_id(
+    request: Request,
     session: Annotated[AsyncSession, Depends(helper_db.session_getter)],
     product: Product = Depends(get_product_by_id),
 ) -> None:
+
+    permission_codes = request.state.user.get("permission")
+    await check_permission(
+        permission_codes=permission_codes,
+        product_code="product.delete",
+    )
     await session.delete(product)
     await session.commit()
